requirements.txt
# Interface do usuário
gradio==3.50.2
pydantic>=1.10.8,<2.0.0

# Processamento de documentos
langchain>=0.0.325
langchain-community>=0.0.13

# Extração de texto
python-docx>=0.8.11
pypdf>=3.15.1

# Banco de dados vetorial
chromadb>=0.4.18

# Embeddings e modelos
sentence-transformers>=2.2.2
huggingface-hub>=0.19.4
openai>=1.6.0
tiktoken>=0.5.2

# Ambiente
python-dotenv>=1.0.0

# Compatibilidade Python 3.13
setuptools>=69.0.2
audioop-lts>=0.0.2 